Update frontend for new data load (#3)
* Commit data generator

* Correct actions path

* Better log which env is not actually set

* Clean up if check

* Add use of github token

* Attempt to upload data in format usable by pages

* remove env for testing

* Add the frontend dashboard

* Commit the cron timer

* Change owner to not conflict with actions var of same name

* Load data from url, refactor load from gui component

* Correct upload location
