+++ .github/workflows/generate-dashboard-data.yaml
@@ -42,22 +42,34 @@ jobs:
         with:
           name: workflow-run-main
           path: ci_dashboard/data_query/workflow_runs.json
-  # generate-angular:
-  #   runs-on: ubuntu-latest
-  #   steps:
-  #   - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
-  #   - uses: actions/setup-node@v4
-  #     with:
-  #       node-version: 20
-  #       cache: 'npm'
-  #       cache-dependency-path: ci_dashboard/frontend/package-lock.json
-  #   - name: Install dependencies
-  #     run: npm ci
-  #   - name: Run Tests
-  #     run: npm test
-  #   - name: Build app
-  #     run: npm run build
+  generate-angular:
+    runs-on: ubuntu-latest
 
+    steps:
+    - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
+    - uses: actions/setup-node@v4
+      with:
+        node-version: 20
+        cache: 'npm'
+        cache-dependency-path: ci_dashboard/frontend/package-lock.json
+    - name: Install dependencies
+      working-directory: ci_dashboard/frontend
+      run: npm ci
+    # - name: Run Tests
+    #   run: npm test
+    - name: Build app
+      working-directory: ci_dashboard/frontend
+      run: npm run build
+    - name: Move files
+      working-directory: ci_dashboard/frontend
+      run: |
+        mv dist/ci-dashboard/3rdpartylicenses.txt dist/ci-dashboard/browser
+        ls -la dist/ci-dashboard
+    - name: Upload artifacts
+      uses: actions/upload-artifact@4cec3d8aa04e39d1a68397de0c4cd6fb9dce8ec1 # ratchet:actions/upload-artifact@v4
+      with:
+        name: frontend
+        path: ci_dashboard/frontend/dist/ci-dashboard/browser
   upload-site:
     if: github.event_name == 'schedule'
     needs: generate-json
@@ -78,12 +90,18 @@ jobs:
           path: data
           pattern: workflow-run-*
           merge-multiple: true
+      - name: Download frontend files
+        uses: actions/download-artifact@cc203385981b70ca67e1cc392babf9cc229d5806 # ratchet:actions/download-artifact@v4
+        with:
+          name: frontend
       - name: Arrange site
         run: |
           ls -la
           mkdir final_site
           mkdir final_site/data
           cp data/* final_site/data/
+          cp frontend/* final_site
+          ls -la final_site/
           ls -la final_site/data/
       - name: Upload static files as artifact
         id: upload
