+++ ci_dashboard/frontend/src/navbar/navbar.component.scss
@@ -22,3 +22,7 @@
   .spacer{
     flex: 1 1 auto;
   }
+
+  .loading-text {
+    font-size: small;
+}
