+++ ci_dashboard/frontend/src/workflow-entry/workflow-entry.component.html
